Revert other unrelated changes
